+++ Program/CCode/include/output.h
@@ -3,6 +3,6 @@
 #ifndef OUTPUT_H_INCLUDED
 #define OUTPUT_H_INCLUDED
 
-void output(char filename[], float time[], float tempWat[], float tempPCM[], float eWat[], float ePCM[], float eTot[], struct parameters params, int sizeOfResults);
+void output(char filename[], double time[], double tempWat[], double tempPCM[], double eWat[], double ePCM[], double eTot[], struct parameters params, int sizeOfResults);
 
 #endif // OUTPUT_H_INCLUDED
