+++ Program/CCode/include/verify_output.h
@@ -3,7 +3,7 @@
 #ifndef VERIFY_OUTPUT_H_INCLUDE
 #define VERIFY_OUTPUT_H_INCLUDE
 
-int verify_output(float tempW[], float tempP[], float eW[], float eP[], struct parameters params, int sizeOfResults);
+int verify_output(double tempW[], double tempP[], double eW[], double eP[], struct parameters params, int sizeOfResults);
 
 #endif // VERIFY_OUTPUT_H_INCLUDE
